:truck: rename: change directory name of container
