+++ knutice/src/components/index.ts
@@ -1,6 +1,6 @@
 export { default as FirstSection } from './Main/FirstSection';
 export { default as SecondSection } from './Main/SecondSection';
-export { default as Home } from './Container/Home';
-export { default as HContent } from './Container/HomeContent';
+export { default as Home } from './Common/Home';
+export { default as HContent } from './Common/HomeContent';
 
-export { default as Navbar } from './Navbar/Navbar';
+export { default as Navbar } from './Common/Navbar';
